Remove git version writing info as it's no longer needed

+++ bayeseor/matrices/build_matrices.py
@@ -17,6 +17,7 @@ from .matrix_funcs import\
     idft_array_idft_1d_sh, IDFT_Array_IDFT_2D_ZM_SH
 from ..model.healpix import Healpix
 from ..utils.utils import get_git_version_info
+from .. import __version__
 
 
 """
@@ -1987,16 +1988,15 @@ class BuildMatrices(BuildMatrixTree):
             print('Prior matrix tree archived but not deleted.'
                   ' \nPath to archive:', path_to_old_matrix_stack)
     
-    def write_git_version_info(self):
+    def write_version_info(self):
         """
-        Write git version info to disk.
+        Write version info to disk.
 
         """
-        version_info = get_git_version_info()
-        fp = Path(self.array_save_directory) / 'git-info.txt'
-        with open(fp, 'w') as f:
-            for key in version_info.keys():
-                f.write('{}: {}\n'.format(key, version_info[key]))
+        fp = Path(self.array_save_directory) / 'version.txt'
+        if not fp.exists():
+            with open(fp, 'w') as f:
+                f.write(f'{__version__}\n')
 
     def build_minimum_sufficient_matrix_stack(
             self,
@@ -2039,8 +2039,5 @@ class BuildMatrices(BuildMatrixTree):
                 confirm_deletion = 'y'
             self.delete_old_matrix_stack(dst, confirm_deletion)
 
-        git_file = Path(self.array_save_directory) / 'git-info.txt'
-        if not git_file.exists():
-            print('Writing git information to disk')
-            self.write_git_version_info()
+        self.write_version_info()
         print('Matrix stack complete')
